Fixed FFI long converter problem on some 64-bit archs.
* The converter typedef was a function returning an int which caused
  problems because the compiler inserted a cltq instruction to sign
  extend %eax when the correct value was already in %rax. Changing
  the converter to a long resolves this.

+++ shotgun/lib/subtend/ffi.h
@@ -29,7 +29,7 @@ OBJECT ffi_new_pointer(STATE, void *ptr);
 #define FFI_TYPE_CHARARR 18
 
 typedef void (*nf_stub_ffi)(void);
-typedef int (*nf_converter)();
+typedef long (*nf_converter)();
 
 #define NFUNC_FIELDS 7
